카테고리 crud 빠진 기능 구현 -한 매장에서 카테고리 삭제 시 다른 매장 카테고리 삭제 안되게
-한 매장에서 카테고리 수정 시 다른 매장에 영향 없게

+++ jumun/src/main/java/com/mihak/jumun/category/CategoryController.java
@@ -2,7 +2,7 @@ package com.mihak.jumun.category;
 
 import com.mihak.jumun.category.form.CategoryForm;
 import com.mihak.jumun.entity.Store;
-import com.mihak.jumun.entity.StoreCategory;
+import com.mihak.jumun.entity.StoreAndCategory;
 import com.mihak.jumun.store.StoreService;
 import com.mihak.jumun.storeCategory.SCService;
 import lombok.RequiredArgsConstructor;
@@ -37,19 +37,29 @@ public class CategoryController {
     @PostMapping("/{storeSN}/category/create")
     public String createCate(Model model , @Valid CategoryForm categoryForm, BindingResult bindingResult, @PathVariable String storeSN) {
         Optional<Category> cate = categoryService.findByName(categoryForm.getName());
+        Store store = storeService.findBySerialNumber(storeSN);
         if(bindingResult.hasErrors()){
             return "category/create_cate";
-        }else if(cate.isPresent()){
-            bindingResult.rejectValue("name", "duplicate",
-                    "중복임");
-            return "category/create_cate";
         }
-        categoryService.create(categoryForm);
-        Optional<Category> cate2 = categoryService.findByName(categoryForm.getName());
-        scService.save(storeSN , cate2.get().getName());
-        System.out.println(storeSN);
+
+        if(cate.isPresent()){
+            Optional<StoreAndCategory> sc = scService.findByStoreAndCategory(store ,cate.get());
+            if(sc.isPresent()){
+                bindingResult.rejectValue("name", "duplicate",
+                        "중복임");
+                return "category/create_cate";
+            }else{
+                scService.save(storeSN, cate.get().getName());
+                return "redirect:/%s/category/list".formatted(storeSN);
+            }
+        }else{
+            categoryService.create(categoryForm);
+            Optional<Category> cate2 = categoryService.findByName(categoryForm.getName());
+            scService.save(storeSN , cate2.get().getName());
+        }
         return "redirect:/%s/category/list".formatted(storeSN);
     }
+
 
     @GetMapping("/{storeSN}/category/list")
     public String showCate(Model model,@PathVariable String storeSN){
@@ -80,28 +90,56 @@ public class CategoryController {
 
     @PostMapping("/{storeSN}/category/modify/{id}")
     public String modify(@PathVariable String storeSN, @Valid CategoryForm categoryForm,BindingResult bindingResult ,Model model , @PathVariable Long id){
-        Category newcate = categoryService.findById(id).get();
-
-        Optional<Category> cate = categoryService.findByName(categoryForm.getName());
+        Store store = storeService.findBySerialNumber(storeSN);
         if(bindingResult.hasErrors()){
-            return "category/cate_modify";
-        }else if(cate.isPresent()){
-            bindingResult.rejectValue("name", "duplicate",
-                    "중복임");
             return "category/cate_modify";
         }
 
-        categoryService.modify(newcate, categoryForm.getName());
+        Optional<Category> beforeModifyCate = categoryService.findById(id);
+        Optional<Category> wantModifyCate = categoryService.findByName(categoryForm.getName());
+
+
+        if(wantModifyCate.isPresent()){
+            Optional<StoreAndCategory> sc = scService.findByStoreAndCategory(store,wantModifyCate.get());
+            if(sc.isPresent()){
+                bindingResult.rejectValue("name", "duplicate",
+                        "중복임");
+
+                return "category/cate_modify";
+            }else {
+                scService.modify(storeSN,beforeModifyCate.get(),wantModifyCate.get());
+                long count = scService.findAllByCategory(id).size();
+                if(count == 0){
+                    categoryService.remove(categoryService.findById(id).get());
+                }
+                return "redirect:/%s/category/list".formatted(storeSN);
+            }
+
+        }else{
+            createCate(model,categoryForm,bindingResult,storeSN);
+            scService.remove(scService.findByStoreAndCategory(store,beforeModifyCate.get()).get());
+        }
+        long count = scService.findAllByCategory(id).size();
+        if(count == 0){
+            categoryService.remove(categoryService.findById(id).get());
+        }
+
         return "redirect:/%s/category/list".formatted(storeSN);
     }
 
     @GetMapping("/{storeSN}/category/delete/{id}")
-    public String delete( @PathVariable("id") Long id, @PathVariable String storeSN) {
-
+    public String delete(@PathVariable("id") Long id, @PathVariable String storeSN) {
+        //해당 스토어와 삭제하고 싶은 카테고리를 가져온다.
+        Store store = storeService.findBySerialNumber(storeSN);
         Category delCate = categoryService.findById(id).get();
-        StoreCategory sc = scService.findByCategory(delCate);
-        scService.remove(sc);
-        categoryService.remove(delCate);
+
+        //StoreAndCategory 테이블에 해당 카테고리를 참조하는 스토어 갯수를 가져온다.
+        long count = scService.findAllByCategory(id).size();
+        scService.remove(scService.findByStoreAndCategory(store,delCate).get());
+
+        if(count == 1){
+            categoryService.remove(delCate);
+        }
 
         return "redirect:/%s/category/list".formatted(storeSN);
     }
